Moved defaults folder
