+++ src/logic/default.ts
@@ -1,8 +1,8 @@
 import Arrow, { IArrow } from "./arrow"
 import Channel, { IChannel } from "./channel"
-import defaultScheme from "./default/data/schemeSet.json"
-import Diagram, { IDiagram } from "./diagram"
+import defaultScheme from "./default/schemeSet.json"
 import { UserComponentType } from "./diagramHandler"
+import Diagram, { IDiagram } from "./diagram"
 import LabelGroup, { ILabelGroup } from "./labelGroup"
 import { ILine, Line } from "./line"
 import { ID } from "./point"
