+++ src/logic/default/diagram.json
