+++ src/logic/default/svgPulse/acquire.json
