+++ src/logic/default/svgPulse/amp.json
