+++ src/logic/default/svgPulse/chirplohi.json
