+++ src/logic/default/svgPulse/halfsine.json
