+++ src/logic/default/svgPulse/index.ts
