+++ src/logic/default/svgPulse/saltirelohi.json
