+++ src/logic/default/svgPulse/trapezium.json
